COMPRESSION=NO, save 169 bytes without compression
This is a challenge to optimize Oiram for size by incrementally
rewriting poorly-compiled (as in pretty large) C functions to assembly.
I am ignoring compression for this challenge as I am not optimizing the
functions for how well they compress. Currently, this saves 169 bytes
compared to the uncompressed version, 9 bytes compared to the
zx7-compressed version, and wastes 86 bytes compared to the
zx0-compressed version, which will be notated by (169,9,-86).

I don't expect this to be merged into the main repository, and I have
yet to test it.